requirements.txt
asgiref==3.3.1
colorama==0.4.4
Django==3.1.4
djangorestframework==3.12.2
Pillow==8.0.1
PyMuPDF==1.18.5
pytz==2020.4
qrcode==6.1
six==1.15.0
sqlparse==0.4.1
